% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_possible_codes.R
\name{cbr_possible_codes}
\alias{cbr_possible_codes}
\title{Get list of possible currencies}
\usage{
cbr_possible_codes(full_output = F)
}
\arguments{
\item{full_output}{By default, FALSE. If TRUE return columns:
\itemize{
  \item Name
  \item EngName
  \item Nominal
  \item ParentCode
  \item ISO_Num_Code
  \item ISO_Char_Code
}}
}
\value{
Data frame, sorted by ISO code with currency names and nominals.
}
\description{
Use this, if you don't know, which ISO code of currency you need.
Or if you don't shure that CBR have this rates or not.
}
