% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_rates.R
\name{cbr_rates}
\alias{cbr_rates}
\title{Get rates of Central Bank of Russian Federation}
\usage{
\code{\link{cbr_possible_codes}}
}
\arguments{
\item{code}{ISO code of currency. By default, USD. You can specify one of sixty possible ISO codes.}

\item{start_date}{By default, 7 days ago.}

\item{end_date}{By default, today.}
}
\value{
Dataframe with date and Currency to 'RUB' to Currency rate.
}
\description{
Get rates: Currency to 'RUB' by dates.
}
\examples{
# How many roubles costs of 1 American dollar during last week?
cbr_rates()

# Dynamic of cost rouble to 1 EUR in 2017
cbr_rates(code = 'EUR', '2017-01-01', '2017-12-31')
}
