% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_rates_with_weekends.R
\name{cbr_rates_with_weekends}
\alias{cbr_rates_with_weekends}
\title{Get CBR rates with filled weekends}
\usage{
\code{\link{cbr_possible_codes}}
}
\arguments{
\item{code}{ISO code of currency. By default, USD. You can specify one of sixty possible ISO codes.}

\item{start_date}{By default, 7 days ago.}

\item{end_date}{By default, today.}
}
\value{
Dataframe with date and Currency to 'RUB' to Currency rate.
Weekends fills last day, when exists CBR rates.
}
\description{
Very useful function, when you have costs or profits on weekend. In \code{\link{cbr_rates}} you get
data without them.
}
\examples{
# For example, today is Monday 2019-04-29. Rates of today will be from Saturday 2019-04-27.
cbr_rates_with_weekends(start_date = '2019-04-29')
}
